+++ src/main/java/com/m1s/m1sserver/Repository/CurriculumScheduleRepository.java
@@ -3,5 +3,5 @@ package com.m1s.m1sserver.Repository;
 import com.m1s.m1sserver.Model.CurriculumSchedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CurriculumScheduleRepository extends JpaRepository<CurriculumSchedule, Integer> {
+public interface CurriculumScheduleRepository extends JpaRepository<CurriculumSchedule, Long> {
 }
