+++ src/main/java/com/m1s/m1sserver/Repository/MemberCurriculumRepository.java
@@ -3,5 +3,5 @@ package com.m1s.m1sserver.Repository;
 import com.m1s.m1sserver.Model.MemberCurriculum;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberCurriculumRepository extends JpaRepository<MemberCurriculum, Integer> {
+public interface MemberCurriculumRepository extends JpaRepository<MemberCurriculum, Long> {
 }
