+++ src/main/java/com/m1s/m1sserver/Repository/MemberScheduleRepository.java
@@ -3,5 +3,5 @@ package com.m1s.m1sserver.Repository;
 import com.m1s.m1sserver.Model.MemberSchedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberScheduleRepository extends JpaRepository<MemberSchedule, Integer> {
+public interface MemberScheduleRepository extends JpaRepository<MemberSchedule, Long> {
 }
