+++ src/main/java/com/m1s/m1sserver/Repository/PartyRepository.java
@@ -3,5 +3,5 @@ package com.m1s.m1sserver.Repository;
 import com.m1s.m1sserver.Model.Party;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PartyRepository extends JpaRepository<Party, Integer> {
+public interface PartyRepository extends JpaRepository<Party, Long> {
 }
